Replace command processor with runner
The command processor is ... weird. How it communicates with the rest of
the system, as well as its purpose are pretty byzantine. So let's get
rid of it. This replaces the command processor with the `Runner`, which
has a similar purpose, but with a much more well defined API. Because
the runner needs to be able to "background" test runs, that is not run
them in the context of where they were initiated, it bundles a context
on which to run the tests.